Markdown.js
Ive started the markdown javascript. I believe Ive added the correct functions and set up for how the readme should be presented. However Im having trouble implementing some of the license information. Im not sure what to add into the if statements for the badge and license link. Also I need to connect the index javascript with the markdown javacript using the dollar sign notation.