Functions, Amrkdown, Package
I've cleaned up pretty much everything else that needed to be done. I accessed the recommeded readme layout and found the license badge... finally. I added some titles to the package,json to make it more descriptive based on the projetc. I also added the functions to the index file. These are not functioning right but this is where Im currently at. The functions seem to be producing the questions twice. Finally I cleaned up some minor things like layout in the README.